Refactored Baker, Dialogue, and Inventory
Created a GameStateTracker script meant to be aware of global happenings in the game via events. Fully refactored the Baker script to rely on the new game state script. Removed any 'HasItem' states since those should be exclusively tracked in the InventoryDataManager to avoid mismatches. Removed and replaced singletons with static methods and properties where applicable.